+++ Libft/Makefile
@@ -79,6 +79,7 @@ FILES = ft_isalpha\
 		ft_strjoin_free_both\
 		ft_free_null\
 		ft_free_arr\
+		ft_isstrws\
 		ft_printf/ft_printf\
 		ft_printf/ft_putchar_pf\
 		ft_printf/ft_putnbr_base_p\
