+++ Libft/ft_isstrws.c
@@ -1,29 +1,27 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   free_utils.c                                       :+:      :+:    :+:   */
+/*   ft_isstrws.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: pgouasmi <pgouasmi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/22 19:01:03 by pgouasmi          #+#    #+#             */
-/*   Updated: 2023/11/23 14:16:39 by pgouasmi         ###   ########.fr       */
+/*   Created: 2023/11/23 13:16:12 by pgouasmi          #+#    #+#             */
+/*   Updated: 2023/11/23 13:16:53 by pgouasmi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cub3D.h"
+#include "../Includes/libft.h"
 
-void	free_maplst(t_dlst *lst)
+int	ft_isstrws(char *str)
 {
-	t_dlst	*temp;
+	size_t	i;
 
-	temp = lst;
-	while (lst)
+	i = 0;
+	while (str[i])
 	{
-		if (lst->content)
-			free(lst->content);
-		temp = lst;
-		lst = lst->next;
-		free(temp);
+		if (!ft_isws(str[i]))
+			return (0);
+		i++;
 	}
-	lst = NULL;
+	return (1);
 }
