+++ Sources/Exit/exit_prog.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   exit_prog.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: chonorat <chonorat@student.42lyon.fr>      +#+  +:+       +#+        */
+/*   By: pgouasmi <pgouasmi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/21 14:54:25 by chonorat          #+#    #+#             */
-/*   Updated: 2023/11/21 14:55:46 by chonorat         ###   ########.fr       */
+/*   Updated: 2023/11/23 13:22:31 by pgouasmi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
