+++ Sources/Parsing/get_map.c
@@ -6,44 +6,57 @@
 /*   By: pgouasmi <pgouasmi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/22 17:57:34 by pgouasmi          #+#    #+#             */
-/*   Updated: 2023/11/22 19:48:06 by pgouasmi         ###   ########.fr       */
+/*   Updated: 2023/11/23 14:41:02 by pgouasmi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3D.h"
 
-int	add_to_maplst(t_maplst **lst, char *str)
+int	initialize_dlst_content(t_dlst *new, char *str)
 {
-	t_maplst	*new;
-	t_maplst	*temp;
-	char		*trim;
+	char	*trim;
 
-	new = malloc(sizeof(t_maplst));
-	if (!new)
-		return (1);
 	trim = ft_strtrim(str, "\n");
 	if (!trim)
 		return (1);
 	new->content = ft_strdup(trim);
 	if (!new->content)
 		return (free(new), free(trim), 1);
+	free(trim);
+	return (0);
+}
+
+int	add_to_maplst(t_dlst **lst, char *str)
+{
+	t_dlst		*new;
+	t_dlst		*temp;
+
+	new = malloc(sizeof(t_dlst));
+	if (!new)
+		return (1);
 	new->next = NULL;
 	if (!*lst)
+	{
 		*lst = new;
+		new->prev = NULL;
+	}
 	else
 	{
 		temp = *lst;
 		while (temp->next)
 			temp = temp->next;
 		temp->next = new;
+		new->prev = temp;
 	}
-	return (free(trim), 0);
+	if (initialize_dlst_content(new, str))
+		return (free(new), 1);
+	return (0);
 }
 
-size_t	maplst_size(t_maplst *lst)
+size_t	maplst_size(t_dlst *lst)
 {
-	size_t		count;
-	t_maplst	*temp;
+	size_t	count;
+	t_dlst	*temp;
 
 	count = 0;
 	temp = lst;
@@ -55,11 +68,11 @@ size_t	maplst_size(t_maplst *lst)
 	return (count);
 }
 
-void	lst_to_arr(t_parser *parser, t_maplst *lst)
+void	lst_to_arr(t_parser *parser, t_dlst *lst)
 {
-	size_t		size;
-	size_t		index;
-	t_maplst	*temp;
+	size_t	size;
+	size_t	index;
+	t_dlst	*temp;
 
 	size = maplst_size(lst);
 	parser->map = malloc(sizeof(char *) * size + 1);
@@ -81,25 +94,97 @@ void	lst_to_arr(t_parser *parser, t_maplst *lst)
 	free_maplst(lst);
 }
 
-void	get_map(t_parser *parser)
+void	parse_before_map(t_parser *parser, t_dlst *lst)
 {
-	t_maplst	*map_list;
-
 	while (1)
 	{
 		parser->line = get_next_line(parser->fd);
 		if (!parser->line)
 			return (free_parser(parser), print_error(PARSING, NOMAP), exit(1));
-		if (!str_isws(parser->line))
+		if (!ft_isstrws(parser->line))
 			break ;
 		free(parser->line);
 		parser->line = NULL;
 	}
-	map_list = NULL;
-	if (add_to_maplst(&map_list, parser->line))
-		return (free_parser(parser), free_maplst(map_list),
+	lst = NULL;
+	if (add_to_maplst(&lst, parser->line))
+		return (free_parser(parser), free_maplst(lst),
 			print_error(MALLOC, 0), exit(1));
 	free(parser->line);
+}
+
+void	remove_empty_down(t_dlst **lst)
+{
+	t_dlst	*temp;
+	t_dlst	*prev;
+	t_dlst	*next;
+
+	temp = *lst;
+	while (temp->next)
+		temp = temp->next;
+	while (!temp->content[0] || ft_isstrws(temp->content))
+	{
+		if (temp == *lst)
+		{
+			prev = temp->prev;
+			next = NULL;
+			prev->next = next;
+			next->prev = prev;
+			if (temp->content)
+				free(temp->content);
+			free(temp);
+		}
+		else
+		{
+			prev = temp->prev;
+			next = temp->next;
+			prev->next = next;
+			next->prev = prev;
+			if (temp->content)
+				free(temp->content);
+			free(temp);
+		}
+		temp = temp->prev;
+	}
+}
+
+void	remove_empty_up(t_dlst **lst)
+{
+	t_dlst	*temp;
+	t_dlst	*prev;
+	t_dlst	*next;
+
+	temp = *lst;
+	while (!temp->content[0] || ft_isstrws(temp->content))
+	{
+		if (temp == *lst)
+		{
+			temp = temp->next;
+			temp->prev = NULL;
+			prev = temp->prev;
+			*lst = temp;
+			if (prev->content)
+				free(prev->content);
+			free(prev);
+		}
+		else
+		{
+			prev = temp->prev;
+			next = temp->next;
+			prev->next = next;
+			next->prev = prev;
+			if (temp->content)
+				free(temp->content);
+			free(temp);
+		}
+		temp = temp->next;
+	}
+}
+
+void	get_map(t_parser *parser)
+{
+	t_dlst	*map_list;
+
 	while (1)
 	{
 		parser->line = get_next_line(parser->fd);
@@ -109,5 +194,7 @@ void	get_map(t_parser *parser)
 			return (free_parser(parser), free_maplst(map_list),
 				print_error(MALLOC, 0), exit(1));
 	}
+	remove_empty_up(&map_list);
+	remove_empty_down(&map_list);
 	lst_to_arr(parser, map_list);
 }
