+++ Sources/Parsing/map_parsing.c
@@ -1,29 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   free_utils.c                                       :+:      :+:    :+:   */
+/*   map_parsing.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: pgouasmi <pgouasmi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/22 19:01:03 by pgouasmi          #+#    #+#             */
-/*   Updated: 2023/11/23 14:16:39 by pgouasmi         ###   ########.fr       */
+/*   Created: 2023/11/23 13:57:47 by pgouasmi          #+#    #+#             */
+/*   Updated: 2023/11/23 14:02:02 by pgouasmi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3D.h"
 
-void	free_maplst(t_dlst *lst)
+void	map_parser(t_parser *parser)
 {
-	t_dlst	*temp;
-
-	temp = lst;
-	while (lst)
-	{
-		if (lst->content)
-			free(lst->content);
-		temp = lst;
-		lst = lst->next;
-		free(temp);
-	}
-	lst = NULL;
+	(void)parser;
 }
